fix(jscpd): extract shared cosmosProto encoder to kill clone
jscpd flagged a 55-line clone between cosmosStaking.ts and
platforms/react-native/chains/cosmos/tx.ts (varint/field/concat/
encodeString/encodeCoin). Extracted to src/utils/cosmosProto.ts;
both files now import from there. varintForTesting re-export in tx.ts
preserved so existing unit tests need no changes.
